added support for custom animations (#8)
* added support for custom animations

* simplified built-in animators

* PR feedback

* more javadocs

+++ android/app/src/main/java/com/reactnativenavigation/utils/ViewUtils.java
@@ -73,14 +73,14 @@ public class ViewUtils {
         }
     }
 
-    public static float getScreenHeight() {
+    public static int getScreenHeight() {
         WindowManager wm = (WindowManager) NavigationApplication.instance.getSystemService(Context.WINDOW_SERVICE);
         DisplayMetrics metrics = new DisplayMetrics();
         wm.getDefaultDisplay().getMetrics(metrics);
         return metrics.heightPixels;
     }
 
-    public static float getScreenWidth() {
+    public static int getScreenWidth() {
         WindowManager wm = (WindowManager) NavigationApplication.instance.getSystemService(Context.WINDOW_SERVICE);
         DisplayMetrics metrics = new DisplayMetrics();
         wm.getDefaultDisplay().getMetrics(metrics);
